Struct bellman::domain::EvaluationDomain

pub struct EvaluationDomain<E: Engine, G: Group<E>> { /* fields omitted */ }

Methods

impl<E: Engine, G: Group<E>> EvaluationDomain<E, G>

pub fn as_ref(&self) -> &[G]

pub fn as_mut(&mut self) -> &mut [G]

pub fn into_coeffs(self) -> Vec<G>

pub fn from_coeffs(
coeffs: Vec<G>
) -> Result<EvaluationDomain<E, G>, SynthesisError>

pub fn from_coeffs_into_sized(
coeffs: Vec<G>,
size: usize
) -> Result<EvaluationDomain<E, G>, SynthesisError>

pub fn fft(&mut self, worker: &Worker)

pub fn ifft(&mut self, worker: &Worker)

pub fn distribute_powers(&mut self, worker: &Worker, g: E::Fr)

pub fn coset_fft(&mut self, worker: &Worker)

pub fn icoset_fft(&mut self, worker: &Worker)

pub fn z(&self, tau: &E::Fr) -> E::Fr

This evaluates t(tau) for this domain, which is tau^m - 1 for these radix-2 domains.

pub fn divide_by_z_on_coset(&mut self, worker: &Worker)

The target polynomial is the zero polynomial in our evaluation domain, so we must perform division over a coset.

pub fn mul_assign(
&mut self,
worker: &Worker,
other: &EvaluationDomain<E, Scalar<E>>
)

Perform O(n) multiplication of two polynomials in the domain.

pub fn sub_assign(&mut self, worker: &Worker, other: &EvaluationDomain<E, G>)

Perform O(n) subtraction of one polynomial from another in the domain.