bellman
Module multiexp

Structs

DensityTracker

FullDensity

Enums

Exponent

An exponent

Traits

AddAssignFromSource

QueryDensity

Source

A source of bases, like an iterator.

SourceBuilder

An object that builds a source of bases.

Functions

multiexp

Perform multi-exponentiation. The caller is responsible for ensuring the query size is the same as the number of exponents.