Module bellman::pairing::ff

Structs

BitIterator

Enums

LegendreSymbol

PrimeFieldDecodingError
An error that may occur when trying to interpret a PrimeFieldRepr as a PrimeField element.

Traits

Field
This trait represents an element of a field.

PrimeField
This represents an element of a prime field.

PrimeFieldRepr
This trait represents a wrapper around a biginteger which can encode any element of a particular prime field. It is a smart wrapper around a sequence of u64 limbs, least-significant digit first.

ScalarEngine
An "engine" is a collection of types (fields, elliptic curve groups, etc.) with well-defined relationships. Specific relationships (for example, a pairing-friendly curve) can be defined in a subtrait.

SqrtField
This trait represents an element of a field that has a square root operation described for it.

Functions

adc
Calculate a + b + carry, returning the sum and modifying the carry value.

mac_with_carry
Calculate a + (b * c) + carry, returning the least significant digit and setting carry to the most significant digit.

sbb
Calculate a - b - borrow, returning the result and modifying the borrow value.