Trait bellman::pairing::ff::SqrtField

pub trait SqrtField: Field {
    fn legendre(&self) -> LegendreSymbol;

    fn sqrt(&self) -> Option<Self>;
}

This trait represents an element of a field that has a square root operation described for it.

Required methods

fn legendre(&self) -> LegendreSymbol

Returns the Legendre symbol of the field element.

fn sqrt(&self) -> Option<Self>

Returns the square root of the field element, if it is quadratic residue.

Loading content...