Trait bellman::pairing::CurveAffine

pub trait CurveAffine: 'static + Send + Copy + Sync + PartialEq<Self> + Eq + Clone + Debug + Display where
    <Self::Engine as ScalarEngine>::Fr == Self::Scalar,
    <Self::Projective as CurveProjective>::Affine == Self,
    <Self::Projective as CurveProjective>::Scalar == Self::Scalar,
    <Self::Uncompressed as EncodedPoint>::Affine == Self,
    <Self::Compressed as EncodedPoint>::Affine == Self,
    <Self::Pair as CurveAffine>::Pair == Self, {
    type Engine: Engine;
    type Scalar: SqrtField + PrimeField;
    type Base: SqrtField;
    type Projective: CurveProjective;
    type Prepared: 'static + Send + Sync + Clone;
    type Uncompressed: EncodedPoint;
    type Compressed: EncodedPoint;
    type Pair: CurveAffine;
    type PairingResult: Field;
    fn zero() -> Self;

    fn one() -> Self;

    fn is_zero(&self) -> bool;

    fn negate(&mut self);

    fn mul<S>(&self, other: S) -> Self::Projective
    where
        S: Into<<Self::Scalar as PrimeField>::Repr>;

    fn prepare(&self) -> Self::Prepared;

    fn pairing_with(&self, other: &Self::Pair) -> Self::PairingResult;

    fn into_projective(&self) -> Self::Projective;

    fn into_compressed(&self) -> Self::Compressed { ... }

    fn into_uncompressed(&self) -> Self::Uncompressed { ... }
}

Affine representation of an elliptic curve point guaranteed to be in the correct prime order subgroup.

Associated Types

type Engine: Engine

type Scalar: SqrtField + PrimeField

type Base: SqrtField

type Projective: CurveProjective

type Prepared: 'static + Send + Sync + Clone

type Uncompressed: EncodedPoint

type Compressed: EncodedPoint

type Pair: CurveAffine

type PairingResult: Field

Loading content...
Required methods

fn zero() -> Self

Returns the additive identity.

fn one() -> Self

Returns a fixed generator of unknown exponent.

fn is_zero(&self) -> bool

Determines if this point represents the point at infinity; the additive identity.

fn negate(&mut self)

Negates this element.

fn mul<S>(&self, other: S) -> Self::Projective where
S: Into<<Self::Scalar as PrimeField>::Repr>,

Performs scalar multiplication of this element with mixed addition.

fn prepare(&self) -> Self::Prepared

Prepares this element for pairing purposes.

fn pairing_with(&self, other: &Self::Pair) -> Self::PairingResult

Perform a pairing

fn into_projective(&self) -> Self::Projective

Converts this element into its affine representation.

Loading content...
Provided methods

fn into_compressed(&self) -> Self::Compressed

Converts this element into its compressed encoding, so long as it's not the point at infinity.

fn into_uncompressed(&self) -> Self::Uncompressed

Converts this element into its uncompressed encoding, so long as it's not the point at infinity.

Loading content...