Trait bellman::pairing::Engine

pub trait Engine: ScalarEngine where
    <Self::G1 as CurveProjective>::Engine == Self,
    <Self::G1 as CurveProjective>::Base == Self::Fq,
    <Self::G1 as CurveProjective>::Scalar == Self::Fr,
    <Self::G1 as CurveProjective>::Affine == Self::G1Affine,
    <Self::G1Affine as CurveAffine>::Engine == Self,
    <Self::G1Affine as CurveAffine>::Base == Self::Fq,
    <Self::G1Affine as CurveAffine>::Scalar == Self::Fr,
    <Self::G1Affine as CurveAffine>::Projective == Self::G1,
    <Self::G1Affine as CurveAffine>::Pair == Self::G2Affine,
    <Self::G1Affine as CurveAffine>::PairingResult == Self::Fqk,
    <Self::G2 as CurveProjective>::Engine == Self,
    <Self::G2 as CurveProjective>::Base == Self::Fqe,
    <Self::G2 as CurveProjective>::Scalar == Self::Fr,
    <Self::G2 as CurveProjective>::Affine == Self::G2Affine,
    <Self::G2Affine as CurveAffine>::Engine == Self,
    <Self::G2Affine as CurveAffine>::Base == Self::Fqe,
    <Self::G2Affine as CurveAffine>::Scalar == Self::Fr,
    <Self::G2Affine as CurveAffine>::Projective == Self::G2,
    <Self::G2Affine as CurveAffine>::Pair == Self::G1Affine,
    <Self::G2Affine as CurveAffine>::PairingResult == Self::Fqk, {
    type G1: CurveProjective + From<Self::G1Affine>;
    type G1Affine: CurveAffine + From<Self::G1>;
    type G2: CurveProjective + From<Self::G2Affine>;
    type G2Affine: CurveAffine + From<Self::G2>;
    type Fq: SqrtField + PrimeField;
    type Fqe: SqrtField;
    type Fqk: Field;
    fn miller_loop<'a, I>(i: I) -> Self::Fqk
    where
        I: IntoIterator<Item = &'a (&'a <Self::G1Affine as CurveAffine>::Prepared, &'a <Self::G2Affine as CurveAffine>::Prepared)>;

    fn final_exponentiation(&Self::Fqk) -> Option<Self::Fqk>;

    fn pairing<G1, G2>(p: G1, q: G2) -> Self::Fqk
    where
        G1: Into<Self::G1Affine>,
        G2: Into<Self::G2Affine>,
    { ... }
}

An "engine" is a collection of types (fields, elliptic curve groups, etc.) with well-defined relationships. In particular, the G1/G2 curve groups are of prime order r, and are equipped with a bilinear pairing function.

Associated Types

type G1: CurveProjective + From<Self::G1Affine>

The projective representation of an element in G1.

type G1Affine: CurveAffine + From<Self::G1>

The affine representation of an element in G1.

type G2: CurveProjective + From<Self::G2Affine>

The projective representation of an element in G2.

type G2Affine: CurveAffine + From<Self::G2>

The affine representation of an element in G2.

type Fq: SqrtField + PrimeField

The base field that hosts G1.

type Fqe: SqrtField

The extension field that hosts G2.

type Fqk: Field

The extension field that hosts the target group of the pairing.

Loading content...
Required methods

fn miller_loop<'a, I>(i: I) -> Self::Fqk where
I: IntoIterator<Item = &'a (&'a <Self::G1Affine as CurveAffine>::Prepared, &'a <Self::G2Affine as CurveAffine>::Prepared)>,

Perform a miller loop with some number of (G1, G2) pairs.

fn final_exponentiation(&Self::Fqk) -> Option<Self::Fqk>

Perform final exponentiation of the result of a miller loop.

Loading content...
Provided methods

fn pairing<G1, G2>(p: G1, q: G2) -> Self::Fqk where
G1: Into<Self::G1Affine>,
G2: Into<Self::G2Affine>,

Performs a complete pairing operation (p, q).

Loading content...