Crate bellman

Modules

domain
This module contains an EvaluationDomain abstraction for performing various kinds of polynomial arithmetic on top of the scalar field.

groth16

pairing

Structs

LinearCombination
This represents a linear combination of some variables, with coefficients in the scalar field of a pairing-friendly elliptic curve group.

Namespace
This is a "namespaced" constraint system which borrows a constraint system (pushing a namespace context) and, when dropped, pops out of the namespace context.

Variable
Represents a variable in our constraint system.

Enums

Index
Represents the index of either an input variable or auxillary variable.

SynthesisError
This is an error that could occur during circuit synthesis contexts, such as CRS generation, proving or verification.

Traits

Circuit
Computations are expressed in terms of arithmetic circuits, in particular rank-1 quadratic constraint systems. The Circuit trait represents a circuit that can be synthesized. The synthesize method is called during CRS generation and during proving.

ConstraintSystem
Represents a constraint system which can have new variables allocated and constrains between them formed.