Trait bellman::ConstraintSystem

pub trait ConstraintSystem<E: Engine>: Sized {
    type Root: ConstraintSystem<E>;
    fn alloc<F, A, AR>(
        &mut self, 
        annotation: A, 
        f: F
    ) -> Result<Variable, SynthesisError>
    where
        F: FnOnce() -> Result<E::Fr, SynthesisError>,
        A: FnOnce() -> AR,
        AR: Into<String>;

    fn alloc_input<F, A, AR>(
        &mut self, 
        annotation: A, 
        f: F
    ) -> Result<Variable, SynthesisError>
    where
        F: FnOnce() -> Result<E::Fr, SynthesisError>,
        A: FnOnce() -> AR,
        AR: Into<String>;

    fn enforce<A, AR, LA, LB, LC>(&mut self, annotation: A, a: LA, b: LB, c: LC)
    where
        A: FnOnce() -> AR,
        AR: Into<String>,
        LA: FnOnce(LinearCombination<E>) -> LinearCombination<E>,
        LB: FnOnce(LinearCombination<E>) -> LinearCombination<E>,
        LC: FnOnce(LinearCombination<E>) -> LinearCombination<E>;

    fn push_namespace<NR, N>(&mut self, name_fn: N)
    where
        NR: Into<String>,
        N: FnOnce() -> NR;

    fn pop_namespace(&mut self);

    fn get_root(&mut self) -> &mut Self::Root;

    fn one() -> Variable { ... }

    fn namespace<'a, NR, N>(
        &'a mut self, 
        name_fn: N
    ) -> Namespace<'a, E, Self::Root>
    where
        NR: Into<String>,
        N: FnOnce() -> NR,
    { ... }
}

Represents a constraint system which can have new variables allocated and constrains between them formed.

Associated Types

type Root: ConstraintSystem<E>

Represents the type of the "root" of this constraint system so that nested namespaces can minimize indirection.

Loading content...
Required methods

fn alloc<F, A, AR>(
&mut self,
annotation: A,
f: F
) -> Result<Variable, SynthesisError> where
F: FnOnce() -> Result<E::Fr, SynthesisError>,
A: FnOnce() -> AR,
AR: Into<String>,

Allocate a private variable in the constraint system. The provided function is used to determine the assignment of the variable. The given annotation function is invoked in testing contexts in order to derive a unique name for this variable in the current namespace.

fn alloc_input<F, A, AR>(
&mut self,
annotation: A,
f: F
) -> Result<Variable, SynthesisError> where
F: FnOnce() -> Result<E::Fr, SynthesisError>,
A: FnOnce() -> AR,
AR: Into<String>,

Allocate a public variable in the constraint system. The provided function is used to determine the assignment of the variable.

fn enforce<A, AR, LA, LB, LC>(&mut self, annotation: A, a: LA, b: LB, c: LC) where
A: FnOnce() -> AR,
AR: Into<String>,
LA: FnOnce(LinearCombination<E>) -> LinearCombination<E>,
LB: FnOnce(LinearCombination<E>) -> LinearCombination<E>,
LC: FnOnce(LinearCombination<E>) -> LinearCombination<E>,

Enforce that A * B = C. The annotation function is invoked in testing contexts in order to derive a unique name for the constraint in the current namespace.

fn push_namespace<NR, N>(&mut self, name_fn: N) where
NR: Into<String>,
N: FnOnce() -> NR,

Create a new (sub)namespace and enter into it. Not intended for downstream use; use namespace instead.

fn pop_namespace(&mut self)

Exit out of the existing namespace. Not intended for downstream use; use namespace instead.

fn get_root(&mut self) -> &mut Self::Root

Gets the "root" constraint system, bypassing the namespacing. Not intended for downstream use; use namespace instead.

Loading content...
Provided methods

fn one() -> Variable

Return the "one" input variable

fn namespace<'a, NR, N>(
&'a mut self,
name_fn: N
) -> Namespace<'a, E, Self::Root> where
NR: Into<String>,
N: FnOnce() -> NR,

Begin a namespace for this constraint system.

Loading content...